ai-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.12.0
numpy==1.26.4
scikit-learn==1.4.0
elasticsearch==8.11.0
python-dotenv==1.0.0
requests==2.31.0